+++ docs/conf.py
@@ -56,7 +56,7 @@ copyright = '2010 by Thomas Johansson, James Crasta'
 # The short X.Y version.
 version = '2.0.1'
 # The full version, including alpha/beta/rc tags.
-release = '2.0.1dev'
+release = '2.0.1'
 
 
 # There are two options for replacing |today|: either, you set today to some